refactor & fix all ts errors
